requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
codecov==2.1.12
colorama==0.4.6
coverage==7.0.2
fpl==0.6.34
frozenlist==1.3.3
idna==3.4
iniconfig==1.1.1
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
packaging==22.0
pluggy==1.0.0
PTable==0.9.2
pytest==7.2.0
pytest-aiohttp==1.0.4
pytest-asyncio==0.20.3
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dotenv==0.21.0
requests==2.28.1
types-requests==2.28.11.7
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
